% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthology.R
\name{getGeneOrthos}
\alias{getGeneOrthos}
\title{get Gene Orthologs}
\usage{
getGeneOrthos(input_org, output_org, gene)
}
\arguments{
\item{input_org}{Species you are starting from. This should be in the latin name format. i.e. Human, Homo Sapien should be "hsapiens"}

\item{output_org}{Species you want to know the orthologs in. Same format as input_org.}

\item{gene}{Ensembl ID of gene in the starting species.}
}
\value{
data frame with all one-to-one orthologs between specified species
}
\description{
Get the one-to-one ortholog for a specific gene between a given species and a desired species
}
